requirements/base.txt
Django~=3.2.0
django-allauth~=0.44.0
django-crispy-forms~=1.11.0
django-debug-toolbar~=3.2.0
whitenoise~=5.2.0
django-environ~=0.4.0
psycopg2~=2.8.0